Mark.6 made
added updater and clean path
