--- release/MGTOOL_5/data/read.txt
@@ -1,1 +1,0 @@
-this dir is for savefile
