--- release/MGTOOL_5/data/today_history.json
@@ -1,109 +1,0 @@
-{
-    "init": [
-        {
-            "user": "User",
-            "time": "2021-07-16 10:31:44 AM",
-            "type": "init",
-            "isSucess": 1
-        },
-        {
-            "user": "User",
-            "time": "2021-07-16 10:32:10 AM",
-            "type": "init",
-            "isSucess": 1
-        }
-    ],
-    "push": [],
-    "show": [
-        {
-            "user": "User",
-            "time": "2021-07-16 10:31:44 AM",
-            "type": "show",
-            "isSucess": 1,
-            "dates": [
-                "2021-07-16"
-            ]
-        },
-        {
-            "user": "User",
-            "time": "2021-07-16 10:31:50 AM",
-            "type": "show",
-            "isSucess": 1,
-            "dates": [
-                "2021-07-16"
-            ]
-        },
-        {
-            "user": "User",
-            "time": "2021-07-16 10:31:51 AM",
-            "type": "show",
-            "isSucess": 1,
-            "dates": [
-                "2021-07-16"
-            ]
-        },
-        {
-            "user": "User",
-            "time": "2021-07-16 10:32:10 AM",
-            "type": "show",
-            "isSucess": 1,
-            "dates": [
-                "2021-07-16"
-            ]
-        },
-        {
-            "user": "User",
-            "time": "2021-07-16 10:32:12 AM",
-            "type": "show",
-            "isSucess": 1,
-            "dates": [
-                "2021-07-14"
-            ]
-        },
-        {
-            "user": "User",
-            "time": "2021-07-16 10:32:12 AM",
-            "type": "show",
-            "isSucess": 1,
-            "dates": [
-                "2021-07-15"
-            ]
-        },
-        {
-            "user": "User",
-            "time": "2021-07-16 10:32:12 AM",
-            "type": "show",
-            "isSucess": 1,
-            "dates": [
-                "2021-07-16"
-            ]
-        },
-        {
-            "user": "User",
-            "time": "2021-07-16 10:32:14 AM",
-            "type": "show",
-            "isSucess": 1,
-            "dates": [
-                "2021-07-16"
-            ]
-        },
-        {
-            "user": "User",
-            "time": "2021-07-16 10:32:22 AM",
-            "type": "show",
-            "isSucess": 1,
-            "dates": [
-                "2021-07-16"
-            ]
-        }
-    ],
-    "built": [
-        {
-            "user": "User",
-            "time": "2021-07-16 10:31:44 AM",
-            "type": "built",
-            "isSucess": 1
-        }
-    ],
-    "fail": []
-}
