--- release/MGTOOL_5/data/today_result.json
@@ -1,64 +1,0 @@
-{
-    "cumulative": 49,
-    "typecode": "TR",
-    "\ub178\ud2b8\ubd81": [
-        {
-            "date": "2021-07-14",
-            "where": "\ub178\ud2b8\ubd81",
-            "\ub0a8\uc131": 11,
-            "\uc5ec\uc131": 3
-        },
-        {
-            "date": "2021-07-15",
-            "where": "\ub178\ud2b8\ubd81",
-            "\ub0a8\uc131": 6,
-            "\uc5ec\uc131": 2
-        },
-        {
-            "date": "2021-07-16",
-            "where": "\ub178\ud2b8\ubd81",
-            "\ub0a8\uc131": 2,
-            "\uc5ec\uc131": 2
-        }
-    ],
-    "\ud504\ub9b0\ud2b8": [
-        {
-            "date": "2021-07-14",
-            "where": "\ud504\ub9b0\ud2b8",
-            "\ub0a8\uc131": 3,
-            "\uc5ec\uc131": 4
-        },
-        {
-            "date": "2021-07-15",
-            "where": "\ud504\ub9b0\ud2b8",
-            "\ub0a8\uc131": 5,
-            "\uc5ec\uc131": 3
-        },
-        {
-            "date": "2021-07-16",
-            "where": "\ud504\ub9b0\ud2b8",
-            "\ub0a8\uc131": 0,
-            "\uc5ec\uc131": 0
-        }
-    ],
-    "\uad00\ub0b4\uc5f4\ub78c": [
-        {
-            "date": "2021-07-14",
-            "where": "\uad00\ub0b4\uc5f4\ub78c",
-            "\ub0a8\uc131": 2,
-            "\uc5ec\uc131": 4
-        },
-        {
-            "date": "2021-07-15",
-            "where": "\uad00\ub0b4\uc5f4\ub78c",
-            "\ub0a8\uc131": 1,
-            "\uc5ec\uc131": 2
-        },
-        {
-            "date": "2021-07-16",
-            "where": "\uad00\ub0b4\uc5f4\ub78c",
-            "\ub0a8\uc131": 0,
-            "\uc5ec\uc131": 0
-        }
-    ]
-}
