--- release/MGTOOL_5/version.txt
@@ -1,1 +1,0 @@
-MK.5
