+++ release/MK.6/data/read.txt
@@ -1,0 +1,1 @@
+this dir is for savefile
