+++ release/MK.6/version.txt
@@ -1,0 +1,2 @@
+MK.6
+MK.6,MK.5,MK.4,MK.3,MK.2
